+++ src/pages/ProblemDetailPage.tsx
@@ -81,6 +81,7 @@ import AnimationVisualization from '../components/visualizations/AnimationVisual
 import AnonymousRecursionVisualization from '../components/visualizations/AnonymousRecursionVisualization';
 import AppendRecordVisualization from '../components/visualizations/AppendRecordVisualization';
 import ApplyCallbackVisualization from '../components/visualizations/ApplyCallbackVisualization';
+import ArbitraryPrecisionVisualization from '../components/visualizations/ArbitraryPrecisionVisualization';
 
 export default function ProblemDetailPage() {
   const { slug } = useParams<{ slug: string }>();
@@ -171,6 +172,7 @@ export default function ProblemDetailPage() {
     'anonymous-recursion': <AnonymousRecursionVisualization />,
     'append-record-to-file': <AppendRecordVisualization />,
     'apply-callback-to-array': <ApplyCallbackVisualization />,
+    'arbitrary-precision-integers': <ArbitraryPrecisionVisualization />,
   };
   
   // Find the problem by slug